Move database actions for Baby table and rename Caretaker to Caregiver
